// src/utils/interfaces/IProducts.tsx
export interface IProduct {
    _id:         string;
    name:        string;
    price:       number;
    category:    string;
    stock:       number;
    description: string;
    brand:       string;
    sku:         string;
    tags:        string[];
    createdAt:   Date;
    updatedAt:   Date;
    __v?:        number;
    specialPrice?: number;
}